// apps/web/src/api/generated/model/questionValueExpression.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Lemma API
 * OpenAPI spec version: 0.1.0
 */

export type QuestionValueExpression =
  | {
      schemaVersion: 1;
      type: "literal";
      value: unknown;
    }
  | {
      schemaVersion: 1;
      type: "reference";
      /** @minLength 1 */
      referenceId: string;
    };